+++ build-tools/src/testFixtures/java/org/elasticsearch/gradle/internal/test/TestClasspathUtils.java
@@ -24,7 +24,7 @@ public class TestClasspathUtils {
     public static void setupJarJdkClasspath(File projectRoot) {
         try {
             URL originLocation = TestClasspathUtils.class.getClassLoader()
-                .loadClass("org.elasticsearch.bootstrap.JdkJarHellCheck")
+                .loadClass("org.elasticsearch.jdk.JdkJarHellCheck")
                 .getProtectionDomain()
                 .getCodeSource()
                 .getLocation();
