+++ libs/core/src/test/resources/org/elasticsearch/jdk/duplicate-classes.jar
