+++ libs/core/src/test/resources/org/elasticsearch/jdk/duplicate-xmlbeans-classes.jar
