+++ server/src/test/java/org/elasticsearch/common/LocalTimeOffsetTests.java
@@ -8,7 +8,7 @@
 
 package org.elasticsearch.common;
 
-import org.elasticsearch.bootstrap.JavaVersion;
+import org.elasticsearch.jdk.JavaVersion;
 import org.elasticsearch.common.LocalTimeOffset.Gap;
 import org.elasticsearch.common.LocalTimeOffset.Overlap;
 import org.elasticsearch.common.time.DateFormatter;
